dom/base/test/jsmodules/importmaps/multiple/test_importMap_after_module_load.html

<!DOCTYPE html>
<head>
  <meta charset=utf-8>
  <title>Test a top-level module load before and after an import map</title>
</head>
<body onload='testLoaded()'>
<script src="chrome://mochikit/content/tests/SimpleTest/SimpleTest.js"></script>

<script>var result;</script>

<!--This is a module load before the import map tag-->
<!--It will import "./module_simpleExport.mjs", which makes "module_simpleExport.mjs"-->
<!--as a resolved module.-->
<script src="./module_simpleImport.mjs" type="module">
</script>

<script>
// eslint-disable-next-line no-unused-vars
let gotMsg = false;
let console = Cc["@mozilla.org/consoleservice;1"].getService(Ci.nsIConsoleService);
let listener = {
  QueryInterface: ChromeUtils.generateQI(["nsIConsoleListener"]),
  observe(msg) {
    info("console message:" + msg);
    ok(msg.logLevel == Ci.nsIConsoleMessage.warn, "log level should be 'warn'.");
    console.unregisterListener(this);
    gotMsg = true;
  }
};
console.registerListener(listener);
</script>

<script type="importmap" onerror='importMapError()'>
{
  "imports": {
    "./module_simpleExport.mjs": "./scope1/module_simpleExport.mjs"
  }
}
</script>

<script>
  /* global gotMsg */
  SimpleTest.waitForExplicitFinish();

  let hasError = false;
  // eslint-disable-next-line no-unused-vars
  function importMapError() {
    hasError = true;
  }

  // eslint-disable-next-line no-unused-vars
  function testLoaded() {
    import("./module_simpleExport.mjs").then((ns) => {
      ok(ns.x == 42, 'Check simple imported value result: ' + ns.x);
      ok(!hasError, "onerror of the import map shouldn't be called.");
      ok(gotMsg, "Should have got the console warning.");
    }).catch((e) => {
      ok(false, "throws " + e);
    }).then(() => {
      SimpleTest.finish();
    });
  }
</script>
</body>
